Setup jwt authentication on tasks

+++ apps/server/src/tasks/tasks.resolver.ts
@@ -1,39 +1,54 @@
-import { Args, Int, Mutation, Query, Resolver } from '@nestjs/graphql';
+import { Headers, Logger, Req, Request, UseGuards } from '@nestjs/common';
+import { Args, Context, Int, Mutation, Query, Resolver } from '@nestjs/graphql';
 import { CreateTaskInput } from './dto/create-task.input';
 import { Task } from './models/task.model';
 import { TaskService } from './tasks.service';
+import { GqlAuthGuard } from 'src/auth/gql.guard';
 
 @Resolver()
 export class TasksResolver {
   constructor(private readonly taskService: TaskService) {}
 
   @Query(() => [Task], { name: 'tasks' })
-  tasks() {
-    return this.taskService.findAll();
+  @UseGuards(GqlAuthGuard)
+  tasks(@Context() { req: { user } }) {
+    return this.taskService.findAll(user.id);
   }
 
-  @Query(() => Task, { name: 'task' })
-  task(@Args('id', { type: () => Int }) id: number) {
-    return this.taskService.findOne(id);
-  }
+  //   @Query(() => [Task], { name: 'tasksAdmin' })
+  //   tasksAdmin() {
+  //     return this.taskService.findAllAdmin();
+  //   }
 
   @Mutation(() => Task)
-  createTask(@Args('createTaskInput') createTaskInput: CreateTaskInput) {
-    return this.taskService.create(createTaskInput);
+  @UseGuards(GqlAuthGuard)
+  create(
+    @Args('createTaskInput') { name, date }: CreateTaskInput,
+    @Context() { req: { user } },
+  ) {
+    return this.taskService.create({ name, date, userId: user.id });
   }
 
   @Mutation(() => Task, { name: 'remove' })
-  deleteTask(@Args('id', { type: () => Int }) id: number) {
-    return this.taskService.delete(id);
+  @UseGuards(GqlAuthGuard)
+  deleteTask(
+    @Args('id', { type: () => Int }) id: number,
+    @Context() { req: { user } },
+  ) {
+    return this.taskService.delete({ id, userId: user.id });
   }
 
   @Mutation(() => Task, { name: 'toggleComplete' })
-  toggleComplete(@Args('id', { type: () => Int }) id: number) {
-    return this.taskService.toggleComplete(id);
+  @UseGuards(GqlAuthGuard)
+  toggleComplete(
+    @Args('id', { type: () => Int }) id: number,
+    @Context() { req: { user } },
+  ) {
+    return this.taskService.toggleComplete({ id, userId: user.id });
   }
 
-  @Mutation(() => Task, { name: 'changeDate' })
-  changeDate(@Args('date', { type: () => Date }) date: Date) {
-    // return this.taskService.toggleComplete({ id, date });
-  }
+  // @Mutation(() => Task, { name: 'changeDate' })
+  // changeDate(@Args('date', { type: () => Date }) date: Date) {
+  //   // return this.taskService.toggleComplete({ id, date });
+  // }
 }
